feat(data): multi-frame per-slice input + gated continuous z
Replace one-frame-per-plane sampling with a fixed S=20 budget: every
in-bbox plane covered once + uniform-random extra frames (planes repeat).
A per-slice cine makes through-plane motion and contraction amplitude
observable from content (no phase label) — the principled fix for the
starved-Δz fork; full coverage keeps the volume L1 loss valid. Slot 0
stays the (t_target, z_mid) reference; t is extraction-only.

Add gated continuous physical z (continuous_z, default off): jitter each
non-reference slot off its integer plane and extract by 2-plane
interpolation so off-grid inference slices splat correctly. Requires
slice_indices as float32 through collation + the gpu_aug gather upgraded
to interpolate (respiratory already float-safe); exact at integer z, so
the default is numerically unchanged.

Loss, model, splat, confidence untouched. Old checkpoints load and run at
S=20 (no arch change). Budget 12->20. See docs/28.

+++ training/data/composed_dataset.py
@@ -154,7 +154,11 @@ class ComposedDataset(Dataset, ABC):
         if "timesteps" in batch:
             sample["timesteps"] = torch.from_numpy(np.stack(batch["timesteps"]).astype(np.int64))
         if "slice_indices" in batch:
-            sample["slice_indices"] = torch.from_numpy(np.stack(batch["slice_indices"]).astype(np.int64))
+            # float32 (not int64): z may be CONTINUOUS (continuous_z). Re-extraction paths
+            # (respiratory grid_sample, gpu_aug 2-plane blend) interpolate; integer-valued z
+            # is exact, so the discrete-grid pipeline is numerically unchanged. timesteps stays
+            # int64 — cardiac phase is always discrete.
+            sample["slice_indices"] = torch.from_numpy(np.stack(batch["slice_indices"]).astype(np.float32))
         if "geom_masks" in batch:
             sample["geom_masks"] = torch.from_numpy(np.stack(batch["geom_masks"]))
         if "rotations" in batch:
